+++ R/play-sox.R
@@ -4,6 +4,8 @@
 #'
 #' The sound is synthesised using \code{\link{save_wav_sox}}
 #' and saved to a temporary file, which is then played from the R session.
+#' This method is generally slower than \code{\link{play_wav}}
+#' but it provides more features.
 #'
 #' @note
 #' The command-line sound-processing program sox
@@ -13,6 +15,8 @@
 #'
 #' @param x Object to play (see \code{\link{save_wav_sox}} for valid options).
 #' @param ... Further parameters to pass to \code{\link{save_wav_sox}}.
+#'
+#' @seealso \code{\link{play_wav}}
 #'
 #' @export
 play_sox <- function(x, ...) {