+++ NEWS.md
@@ -1,3 +1,11 @@
+# hrep 0.12.1
+
+- Exported `play_wav`.
+
+# hrep 0.12.0
+
+- Implemented `save_wav` and `play_wav`, a more streamlined and faster alternative to `save_wav_sox` and `play_wav_sox`.
+
 # hrep 0.11.1
 
 - Improved documentation for sparse_pi_spectrum, sparse_pc_spectrum, and sparse_fr_spectrum.